pios_flash_posix: refactor flash file init into pios
