--- flight/tests/logfs/FreeRTOS.h
@@ -1,3 +1,0 @@
-#include <stdlib.h>
-#define pvPortMalloc(xSize) (malloc(xSize))
-#define vPortFree(pv) (free(pv))
