+++ flight/tests/logfs/pios.h
@@ -1,10 +1,7 @@
+#include <stdbool.h>
+
 /* PIOS Feature Selection */
 #include "pios_config.h"
-
-#if defined(PIOS_INCLUDE_FREERTOS)
-/* FreeRTOS Includes */
-#include "FreeRTOS.h"
-#endif
 
 #if defined(PIOS_INCLUDE_FLASH)
 #include <pios_flash.h>
