+++ flight/tests/logfs/pios_config.h
@@ -1,2 +1,1 @@
 #define PIOS_INCLUDE_FLASH
-#define PIOS_INCLUDE_FREERTOS
